Added CAN Bridge

+++ Source/Candlelight/control.h
@@ -10,9 +10,9 @@
 #include "buffer.h"
 
 void control_init();
-void control_process(int channel, uint32_t tick_now);
-void control_report_busload(int channel, uint8_t busload_percent);
-bool control_send_debug_mesg(int channel, const char* message);
+void control_process(uint8_t channel, uint32_t tick_now);
+void control_report_busload(uint8_t channel, uint8_t busload_percent);
+bool control_send_debug_mesg(uint8_t channel, const char* message);
 bool control_setup_request (USBD_SetupReqTypedef *req);
 void control_setup_OUT_data();
