+++ Source/led.h
@@ -22,11 +22,11 @@ typedef struct
 bool led_init();
 void led_sleep();
 void led_blink_power_on();
-void led_process(int channel, uint32_t tick_now);
-void led_turn_TX(int channel, bool state);
-void led_blink_identify(int channel, bool blink_on);
-void led_flash_TX(int channel);
-void led_flash_RX(int channel);
+void led_process(uint8_t channel, uint32_t tick_now);
+void led_turn_TX(uint8_t channel, bool state);
+void led_blink_identify(uint8_t channel, bool blink_on);
+void led_flash_TX(uint8_t channel);
+void led_flash_RX(uint8_t channel);
 void led_set_Pwr(bool status);
